add registry.py
